requirements.txt
pillow
coloredlogs
pytest
openai
requests
numpy
flake8
pytest-timeout
black
